Restructured layout so that command loading and --help work
Functionality might need further modifications. A few commands seem
still work in progress.
