+++ dev/__init__.py
@@ -1,0 +1,9 @@
+import os
+
+spackdev_aux_subdir = 'spackdev-aux'
+spackdev_aux_bin_subdir = os.path.join(spackdev_aux_subdir, 'bin')
+spackdev_aux_env_subdir = os.path.join(spackdev_aux_subdir, 'env')
+spackdev_aux_packages_subdir = os.path.join(spackdev_aux_subdir, 'packages')
+spackdev_aux_packages_sd_file = spackdev_aux_packages_subdir + '.sd'
+spackdev_aux_specs_subdir = os.path.join(spackdev_aux_subdir, 'spec-yaml')
+spackdev_aux_tmp_subdir = os.path.join(spackdev_aux_subdir, '.tmp')
